Títle: adding save all service to seat entity. Description: It was added save all service for seat entity

+++ travel-agency/reservation-management/src/main/kotlin/com/co/unitravel/domain/models/Seat.kt
@@ -13,7 +13,7 @@ data class Seat(
 
         var seatClass: SeatClass? = null,
 
-        var price: Int? = null,
+        var price: Long? = null,
 
         var customerId: Long? = null,
