+++ travel-agency/reservation-management/src/main/kotlin/com/co/unitravel/infrastructure/adapters/in/rest/controllers/response/SeatResponse.kt
@@ -13,7 +13,7 @@ data class SeatResponse(
 
         var seatClass: SeatClass? = null,
 
-        var price: Int? = null,
+        var price: Long? = null,
 
         var customerId: Long? = null,
